# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: family-finances API
  version: 0.1.0
  description: >-
    The JSON HTTP API of the family-finances backend. This document is
    hand-written and is the source of truth for the API surface (see
    openapi/README.md). Operations tagged `browser-flow` and marked
    `x-internal: true` are browser-navigation redirects, not part of the
    generated typed client.
servers:
  - url: /
    description: Same-origin — the Go backend serves both the API and the SPA.
...
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      summary: The authenticated user's resolved settings
      description: >-
        Always fully populated — a field never set by the user resolves to
        its hardcoded application default.
      tags: [settings]
      responses:
        "200":
          description: The resolved settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: putSettings
      summary: Update one or more of the authenticated user's settings
      description: >-
        Partial update: only the fields present in the body are changed.
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsUpdate"
      responses:
        "200":
          description: The resolved settings after the update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          description: A field included in the body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/tags:
    get:
      operationId: getTags
      summary: List the caller's tags
      tags: [tags]
      responses:
        "200":
          description: Every tag the caller owns.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: postTags
      summary: Create a tag
      tags: [tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagWrite"
      responses:
        "201":
          description: The created tag, owned by the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The caller already has a tag with this name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tags/{id}:
    patch:
      operationId: patchTag
      summary: Rename a tag
      tags: [tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagWrite"
      responses:
        "200":
          description: The updated tag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The caller already has a tag with this name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteTag
      summary: Delete a tag
      description: >-
        Always allowed — detaches the tag from every entry it was attached
        to rather than being blocked by use, unlike a category or account
        type.
      tags: [tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The tag is deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
    bearerToken:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: No valid session was presented.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: >-
        No such resource — including one that belongs to a different owner,
        which behaves identically to nonexistent (existence is not
        disclosed across owners).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: A field in the request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    UserSettings:
      type: object
      additionalProperties: false
      required: [language, timezone, default_currency, displayed_decimal_places]
      properties:
        language:
          type: string
          enum: [en, de]
        timezone:
          type: string
        default_currency:
          type: string
        displayed_decimal_places:
          type: integer
          minimum: 0
          maximum: 4
          description: >-
            Display-only rounding preference — never affects how an amount
            is stored or edited (see account-entries' fixed 4-decimal-place
            storage).
    UserSettingsUpdate:
      type: object
      additionalProperties: false
      properties:
        language:
          type: string
          enum: [en, de]
        timezone:
          type: string
        default_currency:
          type: string
        displayed_decimal_places:
          type: integer
          minimum: 0
          maximum: 4
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        request_id:
          type: string
...
    Tag:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
    TagWrite:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
security:
  - sessionCookie: []
  - bearerToken: []